Update 0.93
This update fixes a number of issues with Deflaktor's ASM Hacks being applied incorrectly. It also properly checks to ensure the input path and output path are not null, and if there is an error, no longer closes the entire program but lets you go back and fix your mistakes.